# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/mediaTypes": {
            "get": {
                "description": "Returns a list of all media types in the system",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "media-types"
                ],
                "summary": "Get all media types",
                "responses": {
                    "200": {
                        "description": "List of media types",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.MediaType"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/messages.Message"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new media type with provided data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "media-types"
                ],
                "summary": "Create new media type",
                "parameters": [
                    {
                        "description": "Media type data",
                        "name": "mediaType",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MediaTypeInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "id",
                        "schema": {
                            "$ref": "#/definitions/messages.Message"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/messages.Message"
                        }
                    },
                    "500": {
                        "description": "Creation error",
                        "schema": {
                            "$ref": "#/definitions/messages.Message"
                        }
                    }
                }
            }
        },
        "/mediaTypes/{id}": {
            "get": {
                "description": "Returns media type by specified identifier",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "media-types"
                ],
                "summary": "Get media type by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Media Type ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Found media type",
                        "schema": {
                            "$ref": "#/definitions/models.MediaType"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/messages.Message"
                        }
                    },
                    "404": {
                        "description": "Media type not found",
                        "schema": {
                            "$ref": "#/definitions/messages.Message"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates media type by ID with provided data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "media-types"
                ],
                "summary": "Update media type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Media Type ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated media type data",
                        "name": "mediaType",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MediaTypeInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated media type",
                        "schema": {
                            "$ref": "#/definitions/http_m.MediaType"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/messages.Message"
                        }
                    },
                    "404": {
                        "description": "Media type not found",
                        "schema": {
                            "$ref": "#/definitions/messages.Message"
                        }
                    },
                    "500": {
                        "description": "Update error",
                        "schema": {
                            "$ref": "#/definitions/messages.Message"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes media type by specified identifier",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "media-types"
                ],
                "summary": "Delete media type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Media Type ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Media type successfully deleted"
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/messages.Message"
                        }
                    },
                    "404": {
                        "description": "Media type not found",
                        "schema": {
                            "$ref": "#/definitions/messages.Message"
                        }
                    },
                    "500": {
                        "description": "Deletion error",
                        "schema": {
                            "$ref": "#/definitions/messages.Message"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.MediaType": {
            "type": "object",
            "properties": {
                "mediaTypeId": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.MediaTypeInput": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
    }
}